F_T30603 move StatementBulkEditTransformer under core transformers
- adjust use statement
